One Front Door/SourceCode/MerchAngular4/dpsg.merchandising.web/src/app/dispatch/ip-main/ip-main.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { IpMainComponent } from './ip-main.component';

describe('Component: IpMain', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [IpMainComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([IpMainComponent],
      (component: IpMainComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(IpMainComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(IpMainComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-ip-main></app-ip-main>
  `,
  directives: [IpMainComponent]
})
class IpMainComponentTestController {
}

